Moved pages into dedicated folders similar to component's folder.
